db data en relations

+++ Project4.0_C3_CrowdCtrl_Back/Project4.0_C3_CrowdCtrl_Back/Models/EventType.cs
@@ -6,7 +6,7 @@ namespace Project4._0_C3_CrowdCtrl_Back.Models
     {
         public int EventTypeId { get; set; }
         public string Name { get; set; }
-        public int Threshold { get; set; }
+        public double Threshold { get; set; }
         [JsonIgnore]
         public List<Event>? Events { get; set; } = new List<Event>();
     }
